Corrects some compilation warnings

+++ Test/BundleAgent/Bundle/CanonicalBlockTest.cpp
@@ -98,7 +98,7 @@ TEST(CanonicalBlockTest, GetMultipleLenght) {
   ss << rawBlock;
   rawBlock = ss.str();
   uint64_t calSize = CanonicalBlock(rawBlock).getLength();
-  ASSERT_EQ(size1, calSize);
+  ASSERT_EQ(static_cast<uint64_t>(size1), calSize);
   rawBlock = rawBlock.substr(calSize);
   calSize = CanonicalBlock(rawBlock).getLength();
   ASSERT_EQ(size2, calSize);
